+++ apps/web/dao/Comment.php
@@ -1,0 +1,10 @@
+<?php
+namespace WDAO;
+
+class Comment extends \MDAOBASE\DaoBase
+{
+    public function __construct($data)
+    {
+        parent::__construct($data);
+    }
+}
