+++ apps/web/dao/Complaints.php
@@ -1,0 +1,10 @@
+<?php
+namespace WDAO;
+
+class Complaints extends \MDAOBASE\DaoBase
+{
+    public function __construct($data)
+    {
+        parent::__construct($data);
+    }
+}
