+++ apps/web/dao/Favorate.php
@@ -1,0 +1,10 @@
+<?php
+namespace WDAO;
+
+class Favorate extends \MDAOBASE\DaoBase
+{
+    public function __construct($data)
+    {
+        parent::__construct($data);
+    }
+}
