+++ apps/web/dao/Msg.php
@@ -1,0 +1,10 @@
+<?php
+namespace WDAO;
+
+class Msg extends \MDAOBASE\DaoBase
+{
+    public function __construct($data)
+    {
+        parent::__construct($data);
+    }
+}
